+++ src/main/kotlin/net/ossrs/ideasrs/explorer/SrsActions.kt
@@ -23,7 +23,8 @@ class SrsRefreshAction(text: String = SrsBundle.message("settings.refresh.descri
     }
 }
 
-class SrsCreateServerAction : DumbAwareAction(SrsBundle.message("srs.create.server.title"), null, AllIcons.Actions.New) {
+class SrsCreateServerAction :
+    DumbAwareAction(SrsBundle.message("srs.create.server.title"), null, AllIcons.Actions.New) {
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.getRequiredData(LangDataKeys.PROJECT)
         val dialog = SrsCreateServerActionDialog(project)